pythonAPI/requirements.txt
numpy==1.15.3
scipy==1.1.0
Pillow==5.3.0
boto==2.49.0
boto3==1.9.253

